Change config to use livenet instead of testnet
Even though development and testing always should be done on the
testnet, the default config should be set to the livenet (mainnet).

The reason for this is that if someone uses the app and forgets to
change the config, no bitcoins would be lost. But if a developer
forgets to change the config, only testnet coins would be lost.

Therefore, the safest option is to use the live config by default.